Clean up MarchingBody code

+++ Space_Sim/Space_Sim/Assets/Game Components/Celestial Objects/Marching Cubes/3D/MarchingChunk.cs
@@ -14,7 +14,7 @@ public class MarchingChunk
 	private MeshRenderer meshRenderer;
 
 	//Settings
-	private int diameter;
+	private float diameter;
     private float terrainScaler;
 
 	private bool smoothTerrain;
